WORK IN PROGRESS:
Nozzle Tips are now global, this also has some impact on nozzle tip runout calibration. Some explanation is needed.

In the physical world, runout can be introduced by the nozzle and/or the nozzle tip. In some machines (e.g. mine), it’s actually the nozzle (more specifically the delicate spring-loaded nozzle tip holder), that has huge runout, not the sturdy CP40 nozzle tips.

As long as we calibrate a nozzle tip each time it is loaded, all that matters is the runout exhibited at the nozzle tip and we need not care where it actually comes from.

However, in order to save job time, PR #825 introduced the possibility to calibrate only once after homing, assuming that unloading/reloading a tip won't rotate it, so the calibration remains valid.

For machines with homed or spring-loaded C axis, the PR even allows to persist the runout calibration. This means runout calibration will add zero time to jobs.

In OpenPNP 1 each nozzle tip was duplicated for each nozzle, so the calibration data was duplicated too, therefore it could correctly capture the combined nozzle + nozzle tip runout.

With global nozzle tips this is no longer true. A multi-nozzle machine could reload the same nozzle tip on a different nozzle, transferring the wrong calibration along.

This commit adds a per nozzle lookup table for the runout calibration. So the nozzle tip can manage calibration for each nozzle separately, while still sharing all the calibration settings and the GUI.

Global nozzle tips also have impact on how they are calibrated when the machine is homed or how calibrations are reset when the machine configuration changes.

This commit is work in progress, as the homing code is currently unattached.