Like discussed with Jason (earlier in PM) the calibration interface was moved to Nozzle.
All calibration methods are now qualified with the nozzle. This removes an ambiguity for the "unloaded" nozzle tip stand-in, when multiple nozzles are naked.

Reverted the nozzle tip homing, as it is no longer needed. The home() override is now again in the Nozzle and since all calibrations are now properly qualified by the nozzle, multiple nozzles will not reset each other's tip calibrations.