Fix: Advanced Camera Calibration startup race.
Somehow, Advanced Camera Calibration configuration needs to be properly
loaded and committed before the first frame is captured. On startup,
Issues & Solutions is finding issues, and displaying the first one. Some
issues query the camera for dimension, promting a camera capture to get
the frame size. By delaying the Issues & Solutions startup, we can
circumvent that.