Improve the error messages when trying to pick a part from the GUI.
This is about nozzle tips compatible with the selected nozzle and with
the package to be picked. If the nozzle tip on the currentkly selected
nozzle is not compatible, the function tries to load a compatible nozzle
tip, if automatic loading is enabled. Conversely, if automatic nozzle
tip loading is disabled, the user is told to consider switching it on.

This fix handles some complications:

- The automatic nozzle tip loading option was suggested, even though it
was already on, but another condition prevented it.
- A compatible nozzle tip could be already loaded on another nozzle, so
the user is instead told to select the other nozzle instead.
- A part could already be loaded on the selected nozzle, preventing
nozzle tip unloading/loading in the first place.

The method `org.openpnp.spi.NozzleTip.getNozzleWhereLoaded()` is
promoted to the `Nozzle` interface to be made available outside
`ReferenceNozzleTip`. The method was renamed to match
`Nozzle.loadNozzleTip()`.