Add Linux support for Neoden 4 cameras using libneodencam
The Neoden4CamDll interface is renamed to Neoden4CameraDriver be-
cause libneodencam exposes the same interface as the Windows dll.

Neoden4CameraHandler now has a `driver` property that is populated
with the respective native library on initialization. If the OS is
identified as Windows, `NeodenCamera.dll` is loaded. Otherwise we
load `libneodencam.1.so`.